+++ python/nvstrings/nvstrings.py
@@ -2637,3 +2637,33 @@ class nvstrings:
         24
         """
         return pyniNVStrings.n_device_memory(self.m_cptr)
+
+    def code_points(self, results):
+        """
+        Fills the given results array with the UTF-8 code point values
+        for each character of each string. Use the len() method to
+        determine the size of each sub-array of integers.
+
+        Parameters
+        ----------
+        results : GPU memory pointer
+            Memory size must be able to hold at least size()*len()
+            of uint32 values.
+
+        Examples
+        --------
+        >>> import nvstrings
+        >>> from librmm_cffi import librmm as rmm
+        >>> import numpy as np
+        >>> s = nvstrings.to_device(["a","xyz", "éee"])
+        >>> s.len()
+        [1, 3, 3]
+        >>> results = rmm.device_array(7,dtype=np.uint32)
+        >>> s.code_points(results.device_ctypes_pointer.value)
+        >>> print(results.copy_to_host())
+        [   97   120   121   122 50089   101   101]
+
+        """
+        if results is None:
+            raise ValueError("results must be device pointer")
+        pyniNVStrings.n_code_points(self.m_cptr, results)