+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - PR #2802 Add `cudf::repeat()` and `cudf.Series.repeat()`
 - PR #2773 Add Fisher's unbiased kurtosis and skew for Series/DataFrame
 - PR #2748 Parquet Reader: Add option to specify loading of PANDAS index
+- PR #2836 Add nvstrings.code_points method
 - PR #2844 Add Series/DataFrame notnull
 
 ## Improvements